# apnashift-backend/openapi.yaml
openapi: 3.0.3
info:
  title: ApnaShift Backend (MVP)
  description: On-demand shifting API — pickup, mini truck, mini tractor. Auth via Bearer JWT.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, example: false }
        error: { type: string, example: invalid_credentials }
    VehicleType:
      type: string
      enum: [pickup, mini_truck, mini_tractor]
    Coord:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, minimum: -90, maximum: 90, example: 22.7196 }
        lng: { type: number, minimum: -180, maximum: 180, example: 75.8577 }
    AddressPoint:
      allOf:
        - $ref: '#/components/schemas/Coord'
        - type: object
          required: [address]
          properties:
            address: { type: string, example: '12 MG Road, Indore' }
...
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        driver_id: { type: string, format: uuid, nullable: true }
        pickup: { $ref: '#/components/schemas/AddressPoint' }
        drop: { $ref: '#/components/schemas/AddressPoint' }
        vehicle_type: { $ref: '#/components/schemas/VehicleType' }
        helper_needed: { type: boolean }
        item_description: { type: string }
        scheduled_time: { type: string, format: date-time, nullable: true }
        delivered_at: { type: string, format: date-time, nullable: true, description: 'Set on delivered, else null' }
        distance_km: { type: number }
        price_rs: { type: number }
        status:
          type: string
          enum: [pending, accepted, arrived, in_transit, delivered, cancelled]
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /api/bookings:
    post:
      summary: Create booking (user, price computed server-side)
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional retry key (A-Za-z0-9_-, 1-64). Same key + same payload replays 200 with same booking; same key + different payload is 422.
          schema: { type: string, example: order-abc-123 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pickup, drop, vehicle_type]
              properties:
                pickup: { $ref: '#/components/schemas/AddressPoint' }
                drop: { $ref: '#/components/schemas/AddressPoint' }
                vehicle_type: { $ref: '#/components/schemas/VehicleType' }
                helper_needed: { type: boolean, default: false }
                item_description: { type: string }
                scheduled_time: { type: string, format: date-time, description: 'Optional, must be now/future (past date is 400)' }
      responses:
        '201':
          description: Created (status pending)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  booking: { $ref: '#/components/schemas/Booking' }
        '200':
          description: Idempotent replay (same key + same payload, same booking)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  booking: { $ref: '#/components/schemas/Booking' }
        '400':
          description: validation_failed (incl. past scheduled_time) or distance_too_far (>500 km) or invalid_idempotency_key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: idempotency_conflict (same key, different payload)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      summary: Own bookings, paginated (user)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200': { description: OK }
...
  /api/admin/drivers:
    get:
      summary: Drivers by status (admin, paginated)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, verified, review], default: pending } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200': { description: OK }
        '403': { description: forbidden (non-admin) }
...
  /api/admin/bookings:
    get:
      summary: All bookings with filters + totals (admin)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, accepted, arrived, in_transit, delivered, cancelled] } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: city, in: query, schema: { type: string } }
      responses:
        '200': { description: total + revenue_rs + rows }
...